refactor the audio uploader (#5)
* refactor the audio uploader

* refactoring, split up components
