+++ src/components/AudioTimeSlider/AudioTimeSlider.module.css
@@ -1,0 +1,12 @@
+.time-slider {
+  width: 100%;
+}
+
+.control-time {
+  margin: 1em 0 0 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0.5em;
+  width: 140%;
+}
