+++ src/components/MediaControlPanel/MediaControlPanel.module.css
@@ -5,20 +5,6 @@
   flex-direction: column;
 }
 
-.control-button {
-  margin: 1em 0 0 0;
-  padding: 20px;
-  border: 0;
-  border-radius: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.skip-button {
-  padding: 10px;
-}
-
 .control-group {
   display: flex;
   justify-content: center;
@@ -34,17 +20,4 @@ input[type="range"]::-webkit-slider-thumb {
   justify-content: center;
   align-items: center;
   gap: 1em;
-}
-
-.time-slider {
-  width: 100%;
-}
-
-.control-time {
-  margin: 1em 0 0 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 0.5em;
-  width: 140%;
 }
